+++ src/main/java/spring/springserver/domain/member/data/request/ChangeUsernameRequest.java
@@ -1,4 +1,4 @@
-package spring.springserver.domain.auth.data.request;
+package spring.springserver.domain.member.data.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.Email;
