+++ src/main/java/spring/springserver/domain/member/data/request/PasswordResetRequest.java
@@ -1,4 +1,4 @@
-package spring.springserver.domain.auth.data.request;
+package spring.springserver.domain.member.data.request;
 
 import jakarta.validation.constraints.NotBlank;
