+++ src/main/java/spring/springserver/global/config/SecurityConfig.java
@@ -30,7 +30,7 @@ public class SecurityConfig {
 	}
 
 	@Bean
-	public SecurityFilterChain filterChain(HttpSecurity httpSecurity) {
+	public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception{
 
 		httpSecurity
 				.httpBasic(AbstractHttpConfigurer::disable)
